Move pretty printing into modules
