+++ crates/ddl/src/diagnostics.rs
@@ -3,8 +3,7 @@
 use codespan::{ByteIndex, FileId, Span};
 use codespan_reporting::diagnostic::{Diagnostic, Label, Severity};
 
-use crate::core;
-use crate::surface::delaborate;
+use crate::{core, surface};
 
 pub fn field_redeclaration(
     severity: Severity,
@@ -56,10 +55,10 @@ pub fn type_mismatch(
 ) -> Diagnostic {
     let arena = pretty::Arena::new();
 
-    let expected_ty = delaborate::delaborate_term(&core::semantics::readback(expected_ty));
-    let found_ty = delaborate::delaborate_term(&core::semantics::readback(found_ty));
-    let pretty::DocBuilder(_, expected_ty) = expected_ty.doc(&arena);
-    let pretty::DocBuilder(_, found_ty) = found_ty.doc(&arena);
+    let expected_ty = surface::delaborate::delaborate_term(&core::semantics::readback(expected_ty));
+    let found_ty = surface::delaborate::delaborate_term(&core::semantics::readback(found_ty));
+    let pretty::DocBuilder(_, expected_ty) = surface::pretty::pretty_term(&arena, &expected_ty);
+    let pretty::DocBuilder(_, found_ty) = surface::pretty::pretty_term(&arena, &found_ty);
     let expected_ty = expected_ty.pretty(100);
     let found_ty = found_ty.pretty(100);
 
@@ -89,8 +88,8 @@ pub fn universe_mismatch(
 ) -> Diagnostic {
     let arena = pretty::Arena::new();
 
-    let found_ty = delaborate::delaborate_term(&core::semantics::readback(found_ty));
-    let pretty::DocBuilder(_, found_ty) = found_ty.doc(&arena);
+    let found_ty = surface::delaborate::delaborate_term(&core::semantics::readback(found_ty));
+    let pretty::DocBuilder(_, found_ty) = surface::pretty::pretty_term(&arena, &found_ty);
     let found_ty = found_ty.pretty(100);
 
     Diagnostic {
@@ -247,8 +246,8 @@ pub mod error {
     ) -> Diagnostic {
         let arena = pretty::Arena::new();
 
-        let found_ty = delaborate::delaborate_term(&core::semantics::readback(found_ty));
-        let pretty::DocBuilder(_, found_ty) = found_ty.doc(&arena);
+        let found_ty = surface::delaborate::delaborate_term(&core::semantics::readback(found_ty));
+        let pretty::DocBuilder(_, found_ty) = surface::pretty::pretty_term(&arena, &found_ty);
         let found_ty = found_ty.pretty(100);
 
         Diagnostic {
@@ -283,8 +282,8 @@ pub mod error {
     ) -> Diagnostic {
         let arena = pretty::Arena::new();
 
-        let found_ty = delaborate::delaborate_term(&core::semantics::readback(found_ty));
-        let pretty::DocBuilder(_, found_ty) = found_ty.doc(&arena);
+        let found_ty = surface::delaborate::delaborate_term(&core::semantics::readback(found_ty));
+        let pretty::DocBuilder(_, found_ty) = surface::pretty::pretty_term(&arena, &found_ty);
         let found_ty = found_ty.pretty(100);
 
         Diagnostic {
